+++ src/translations/en/cookie-boss.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Craft Cookie boss plugin for Craft CMS 3.x
+ *
+ * Allow your visitors to set there cookie preference.
+ *
+ * @link      www.dutchheight.com
+ * @copyright Copyright (c) 2019 Dutch Height
+ */
+
+/**
+ * Craft Cookie boss en Translation
+ *
+ * Returns an array with the string to be translated (as passed to `Craft::t('craft-cookie-boss', '...')`) as
+ * the key, and the translation as the value.
+ *
+ * http://www.yiiframework.com/doc-2.0/guide-tutorial-i18n.html
+ *
+ * @author    Dutch Height
+ * @package   CookieBoss
+ * @since     1.0.0
+ */
+return [
+
+];